v0.0.6: Event-driven theming, sprite fixes, and shell integration
Features:
- Event-driven reactive theming (command success/fail, bell, focus)
- Sprite animation patches in event handlers
- Shell integration for OSC 133 semantic prompts

Fixes:
- Sprite patch paths now resolve relative to theme directory
- Shell integration compatibility with p10k/oh-my-zsh
- Theme hot reload path handling on macOS

Documentation:
- Troubleshooting guide for shell integration issues
- Updated theming docs with event handler examples

+++ crates/crt-renderer/src/effects/shape.rs
@@ -227,7 +227,7 @@ impl ShapeEffect {
         let height = bounds.height();
         let cx = width / 2.0;
         let cy = height / 2.0;
-        let speed = self.motion_speed as f64;
+        let speed = self.motion_speed;
         let t = self.time * speed;
 
         let (x, y) = match self.motion_type {
